requirements.txt
# MPT-CRM Dependencies

# Web Framework
streamlit>=1.32.0
streamlit-sortables>=0.2.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Email
sendgrid>=6.10.0

# Data Processing
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.0
